refactor: Update DeployDSC contract to return HelperConfig instance
This commit updates the DeployDSC contract to return an instance of the HelperConfig contract along with the DecentralizedStableCoin and DSCEngine instances. This change allows for easier access to the HelperConfig contract for configuring network settings.

Refactor the `run` function in the DeployDSC contract to return `(DecentralizedStableCoin, DSCEngine, HelperConfig)` instead of just `(DecentralizedStableCoin, DSCEngine)`.

No issues or tags are associated with this commit.